server/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
python-dotenv==1.1.1
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
requests==2.32.5
Pillow==10.4.0
